domain name changes: yegor256 -> artipie
