+++ src/main/java/com/artipie/npm/Npm.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package com.yegor256.npm;
+package com.artipie.npm;
 
 import com.yegor256.asto.Storage;
 import io.reactivex.rxjava3.core.Completable;
