+++ src/main/java/com/artipie/npm/NpmPublishJsonToMetaBind.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package com.yegor256.npm;
+package com.artipie.npm;
 
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
