+++ src/test/java/com/artipie/npm/NpmCommandsTest.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package com.yegor256.npm;
+package com.artipie.npm;
 
 import com.yegor256.asto.Storage;
 import io.vertx.reactivex.core.Vertx;
